Day 29: BitwiseAND Solution. 30- Days-of-Code Hackerrank Solution in Java.
This program takes input N, then calculates AND between A&B such that
A<B and B<N. And whatever the value of A AND B is maximum but less than
K, it displays that value. For simplicity uncomment the print line below
and read the Hackerrank description and reun below code, you'll
understand.